Plug in WP_Git_Filesystem to directly browse and edit the static files from a git repo
