+++ packages/playground/data-liberation/bootstrap.php
@@ -81,6 +81,11 @@ require_once __DIR__ . '/src/import/WP_Entity_Iterator_Chain.php';
 require_once __DIR__ . '/src/import/WP_Retry_Frontloading_Iterator.php';
 require_once __DIR__ . '/src/import/WP_Import_HTML_Processor.php';
 require_once __DIR__ . '/src/import/WP_Import_Utils.php';
+
+require_once __DIR__ . '/src/git/WP_Git_Client.php';
+require_once __DIR__ . '/src/git/WP_Git_Pack_Index.php';
+require_once __DIR__ . '/src/git/WP_Git_Filesystem.php';
+
 require_once __DIR__ . '/src/WP_Data_Liberation_HTML_Processor.php';
 require_once __DIR__ . '/src/utf8_decoder.php';
